angle: Fix names of Windows testers.
The angle.json file explicitly says the arch of the platform:
win7-x86-amd, win10-x64-intel and win10-x64-nvidia. It's a better
naming scheme because it allows for both x86 and x64 testers for
the same GPU vendor. Also corrects the triggering to use the
correct x86 for AMD and x64 for Intel/NV.

Confirmed with led that the correct tests are now triggered.

Test: led launch
Bug: angleproject:5114
Change-Id: I5cc0b48fb5f545a6f25c60d92075869013b34c0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2857814
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Jamie Madill <jmadill@chromium.org>