Use the adjust-priority flag when calling get-builder.
Prior to this change, before launching the led jobs, the tester would
manually restore the priority to the builder's default by increasing it
by 10, which assumes implementation details of led. This change switches
to the newly added -adjust-priority flag, which is a supported method
for controlling the priority adjustment when getting an led job.

Change-Id: I1a83888f0856ca75eb2fc481d68c34b761f39f38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2874979
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>