Remove automatic archive step from emscripten_releases recipe
Currently the recipe automatically runs the archive step to upload
binaries after building them. But the cross builder will have 2
archive steps. So instead we will add the archive step explicitly
to the list of build steps in the emscripten-releases repo.
So this change removes the step, and also the property that
controlled it (the linux-test-suite builder will simply not have
an archive step).

Bug: v8:11309
Change-Id: Iacf24355ee16edcb5ae5aad377842e6b664c8aaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2877930
Commit-Queue: Derek Schuff <dschuff@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>