Turn off shard-level retries on mac11.0-blink-rel.
This CL turns off the "retry failed shards" logic for the Mac11
rebaseline builder on tryserver.blink (mac11.0-blink-rel);
we're low on capacity and retrying failed shards makes things
slower and harder to manage.

Really, we should turn off retries for all the builders on
tryserver.blink, because there's no good excuse for tolerating
flaky failures there (the main reason we retry things), but
we don't have a good mechanism in place yet on the Blink team
for addressing this.

Bug: 1199850
Change-Id: I5b45dbb884dd8d17ec06e52c33e804130804308b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2878177
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Reviewed-by: Gregory Guterman <guterman@google.com>