Changes to flakiness module to optimize ease-of-use and readability.
After extensive LED testing, a few changes were needed to improve
readability and appearance of the module itself, and its step
presentation.

This CL included several changes, the major ones being:
- The addition and renaming of an input property. This allows for the
number of tests queried differ across the historical data and current
build data and caters to the common instance in which a more specific
current build test set is needed without the increased runtime of a more
specific historical test set.
- The renaming of the get_test_result_history step to verify_new_tests.
The former is extremely vague, especially in the context of our module.
- Changing the presentation format. This included the logging of the
tests for the current build, the preliminary set of positive tests, and
renaming "associated invocations" to "excluded invocations" to clarify
the purpose of the list.
- Improving docstring consistency.

Bug: 1204163
Change-Id: I36d82acb0eca47f1327c495e9755745f7485a50f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2935096
Commit-Queue: Marco Georgaklis <mgeorgaklis@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>