test_utils: Fetch RDB results one suite at a time.
Right now, we're fetching all results for every suite in a single
`rdb query` invocation. Given the format of the returned results
(ie: a flat list of results), this makes it difficult to pull out and
group each individual test result with its suite.

By calling `rdb query` once for each suite, we can easily group the
results ourselves. This will be useful when we start attaching these
results to each suite's steps.Test object (will be done in a follow-up).

Note that this could be a regression in overall build time since one
invocation for a bunch of tasks is likely faster than a bunch of
invocations for a few tasks each. A few reasons this might be worth the
regression:
- It makes the recipe code simpler. (It'd need some annoying dict
  iterations to group things correctly.)
- This is how the recipe will likely run in the end state of RDB
  integration. Instead of calling `rdb query` for every suite once
  they've all finished, call it once for each suite once that suite's
  collect() step has finished.

Recipe-Nontrivial-Roll: build_limited_scripts_slave
Bug: 1135718
Change-Id: I7fc1273a7ec17c8db2c455efcbef7c78e6dd5af9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2937897
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>