Filter out known-flaky tests from the suites in RDB's retry decision.
The legacy decision that decides which suites to retry after the
"with patch" phase filters out any tests known by FindIt to be flaky.
The RDB-based for the same thing currently does not do the same
filtering. This leads to differences in the two decisions.

By incorporating the results returned by FindIt, we can narrow down
the differences between the two decisions.

Confirmed that this correctly filters out flaky tests in the RDB
decision. eg, this build had the cc_unittests failure
forgiven from the legacy retry decision but not the RDB decision:
https://ci.chromium.org/p/chromium/builders/try/linux_chromium_tsan_rel_ng/876373?

Re-running that with this patch applied results in the same failure
being forgiven for both decisions:
https://ci.chromium.org/raw/build/logs.chromium.org/chromium/led/bpastene_google.com/f66576840ee5ba4ae92d659660f1bd170bb0d4a03bc9447f8768bb4e32deefc8/+/annotations?server=chromium-swarm.appspot.com

Bug: 1135718
Change-Id: I499d82a238cd95585790c7dc2a5719e09b427f5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2911001
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>