test_utils: Include the 'expected' field when fetching RDB results.
crrev.com/c/2981355 accidentally caused this field to be dropped since
it's included by default when no fields are specifically requested, but
is excluded when they are.

So add it back to the results so we can correctly ignore expected
failures again.

Recipe-Nontrivial-Roll: build_limited_scripts_slave
Bug: 1135718
Change-Id: I07e113e34ce48b17aa149e4b1499154a8f3f0bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2986406
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>