test_utils: Wrap local ScriptTests in an rdb invocation.
This re-uses the code paths that wrap LocalIsolatedScriptTests in an
rdb invocation, and applies it to ScriptTests (which are also ran
locally).

These script tests were instrumented to be able to upload their results
to the result-sink in crrev.com/c/2977971. So once this lands, they
should automatically start uploading to RDB.

Bug: 1135718
Change-Id: Ib5b03cf66c4d2c20fccd6053f7d5313c1781a311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2993598
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>