Add ssd builder for dev.
This builder should verify that builds are being done
on available ssds. Initially this will be connected to
a bot without enough storage to build and should fail.
I'll pause the builder and resume it when an ssd has
been attached to the builder to verify that it is
succeeding because of the ssd

Bug:1218948
Change-Id: Ibb6e2cfbae04f1448b1f8aa5c14300f4cb33661e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/2997262
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>