Report the properties that the recipe sees.
The properties displayed in Milo are the properties that the chromium
bootstrapper sees. Generally, the recipe will see a different set of
properties because the bootstrapper will read properties for the
builder from a file and set the properties with those values. In order
to be able to debug properties-related issues, an additional step is
being added that will be run automatically for any build that was
bootstrapped by the chromium bootstrapper. This additional step has a
log containing the properties as the recipe sees them.

Bug: 1210566
Change-Id: Idd71a5fbc5e39f027aff8831d6de8a63fe4652fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3011452
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>