test_utils: Fetch results of ScriptTests from RDB.
Similar to what we're doing for swarming tests, and now for local
script tests.

Note that we can't count on a consistent naming scheme for the
invocations of the local script tests. So for now we have to parse the
stderr of the test for the rdb invocation name and pull that out.

But other than that, the results should be able to be handled the same
as the swarming tests. (eg: We fetch the results the same way, parse
them the same way, etc.)

Bug: 1135718
Change-Id: I16610ceea8c477f1e91f3fbec769b138082c3417
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3011440
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>