Run all local script tests through vpython.
Some of the tests import non-standard modules. eg: see the 'six' error
at the bottom of:
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket.appspot.com/8842248141098896832/+/u/check_network_annotations__with_patch_/stdout

That's likely because it's using the system's six (?), which is old and
lacks the 'ensure_str' method.

Using vpython should make it use the vpython's spec file's six, which
is new enough.

Bug: 1135718
Change-Id: I6f95c16dd8dbce1d27d2abc6ae2aca7425ba3d4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3016433
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>