Remove upload of WebRTC.framework to the chromium-webrtc bucket.
This step started to fail after the recent update to the script to
generate WebRTC.xcframework. Since WebRTC should be built from source
anyway, let's remove this step which archives the build output on
a bucket (not used or supported as far as I know).

Bug: None
Change-Id: I63edeca357d39c5511139db2a48dd64161facfc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3015212
Reviewed-by: Andrey Logvin <landrey@google.com>
Commit-Queue: Mirko Bonadei <mbonadei@chromium.org>