Check out src-internal at the target ref's HEAD for src-internal CLs.
When running against src-internal CLs, a repo_path_map value of
('src-internal', None) indicates that the patch should be applied
against whatever version of src-internal is being checked out by the
gclient config. If the revision for src-internal isn't set, then that
means it will be applied to whatever src-internal's pin is at for src.
This caues problems where a change is based off of a newer revision of
src-internal than that pin and won't apply cleanly to the older version.
In that case, the try builder would fail to apply the patch until a
newer version of src-internal rolls into src.

Using ('src-internal', 'HEAD') instead causes the head of the CL's
target ref to be checked out if the revision isn't otherwise set.

Change-Id: I2d179f48fa968bb13aa06e9260481d4e2d6975f0
Recipe-Nontrivial-Roll: build_limited
Recipe-Nontrivial-Roll: chrome_release
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3024697
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>