Remove android & fuchsia specific logic from local gtests.
All fuchsia and android gtests should(?) be on swarming at this point,
so this should be safe.

And putting tests on swarming is still very much the preferred setup
for new bots. So ideally this support won't be needed anytime in the
future either.

Note that fuchsia and android bots can still specify local gtests, but
this CL will cause them to use the exact same logic that local desktop
gtests use today. It may not work out of the box, so some modernizing
of `api.chromium.runtest` might be needed should anyone add a local
fuchsia/android gtest in the future.

Bug: 1135718
Change-Id: I6e18ff393e962d92845fe8bfdc8c8fbf2b6d7d30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3027699
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>