Update inv_name mapping for local script tests even if they fail.
It seems the run() step for local script tests are expected to raise
an exception upon failure. That makes anything outside of the "finally:"
block not executed when the test exits non-zero.

So this moves the update_inv_name_from_stderr() call inside the
finally block so it's executed even on failure.

eg: See the "No RDB results" warning in:
https://ci.chromium.org/p/chromium/builders/ci/linux-annotator-rel/54422

And see it disappear in this led job:
https://ci.chromium.org/swarming/task/552125cf8d56ad10?server=chromium-swarm.appspot.com

Bug: 1135718
Change-Id: I537d6eff0e71897375efa171fe81b798c996b934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3068003
Reviewed-by: Stephen Martinis <martiniss@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>