Don't build tests for testers that are triggered by a mirrored builder.
Currently, when a try builder runs, it computes the set of tests
requested by the builders and testers that it mirrors and it also
computes the set of tests requested by any testers triggered by its
mirrored builders. Any tests in the second set will be built but not
executed. This provides an undocumented benefit of ensuring that the
tests for these triggered testers continue to compile with the change
under test, but has some drawbacks:
* When source side spec files are changed, we do a full build and
  execute all tests, so if a triggered tester is not on the same builder
  group as one of the mirrors, unrelated changes will cause a full
  build.
* It is easy to add additional compile targets to a try builder without
  realizing it by having a tester be triggered by a mirrored builder
  and/or adding additional tests.

This change causes triggered testers to be ignored for the try builder
codepath. Instead, only targets/tests for the explicitly configured
mirrors will be built and/or executed. If we desire functionality to
ensure that compile targets for triggered testers continue to build,
then we should add support for explicitly configuring a mirrored tester
to not have tests executed.

Change-Id: I42936b334f8a6ee717ce49a578b947e6ca9ec85a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3072058
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>