Allow users to specify different dut_pool
Skylab pipeline was initially designed for lacros, whose tests are always
running in quota pool, the shared dut pool with other CrOS tests.
Now we have NearbyConnection to onboard, which are running in a pool
with special hardware. So add a new field, dut_pool, to allow users
to specify it from chromium side.

BUG=b/194948794
TEST=train

Change-Id: I367f33d4ca2e10960e0c36ab9770fd29a26e5a21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3077464
Commit-Queue: Xinan Lin <linxinan@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>