Move mapping for pinpoint builders into pinpoint recipe.
The pinpoint builder recipe is relying on the current behavior where all
the tests for any testers that are triggered by a mirrored builder will
be built. The pinpoint builder recipe doesn't take advantage of any
other trybot functionality in chromium_tests.

The pinpoint builder recipe also doesn't match the expectations that the
chromium_tests module has for try builders since it is run with and
without a CL to generate test isolates that are then triggered via some
other mechanism to compare results.

To reduce inappropriate coupling between the pinpoint builder and the
trybot functionality, the recipe now contains a mapping from the
pinpoint builder to the perf builder and it now does a lookup for the CI
builder without using the try_db. The returned builder config will cause
tests for all triggered testers to be built.

Bug: 1237605
Change-Id: I012181bb721361fc8892bc18c19f2ec823b9017a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3082899
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ryan Heise <heiserya@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ryan Heise <heiserya@google.com>