test_utils: Track invalid results alongside tests' RDB result.
In a world where RDB is the only source of truth for test results, a
test may crash at-startup and fail to upload any results to RDB. A naive
recipe might then query RDB in that case, see no failed tests, and
consider the test as having passed.

To avoid such errors, this CL will make the recipe record for every test
in steps.py whether that test exited with a non-zero exit code. We
combine that single boolean with the results from RDB to ensure that no
matter what a test uploads to RDB, it will still be considered a
failure if its exit code is non-zero.

Recipe-Nontrivial-Roll: build_limited
Recipe-Nontrivial-Roll: chrome_release
Bug: 1135718
Change-Id: Ia5f28cac194ff8e06a8f3d5ebc51dc4f92888383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3015244
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>