test_utils: Use steps.Test test name when creating RDB results.
Currently the recipe will parse the results in RDB for the suite name
of a test. But this can break when RDB has no results for that test, eg:
https://ci.chromium.org/p/chromium/builders/try/win10_chromium_x64_rel_ng/941891?
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket.appspot.com/8839179567272176176/+/u/RECIPE_CRASH__Uncaught_exception_/logging

The canonical_name field in steps.Test is the same thing that gets
tagged in RDB results, so we can just pass in the steps.Test field
directly.

Bug: 1135718
Change-Id: I372ae1262a6115cf47ec983f94d78a67e7d2898b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3089517
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>