Do not build/run tests for non-mirrored testers.
To simplify the declaration of a number of existing compile-only trybots
that mirror only a builder and expect to build the tests for all of the
builder's triggered testers, the `include_all_triggered_testers`
attribute was added to TrySpec. This prevents the maintenance burden of
having to keep trybot definitions in sync with the set of triggered
testers while also being explicit about relying on the behavior.

The default TrySpec created when looking up a builder now sets this
field, which takes care of the CI builders needing to read the test
specs for their triggered testers. As part of this process, specific
terminology was introduced for describing the builders that would in one
way or another be referenced by a builder config.

With the removal of the behavior of implicitly building targets for
non-mirrored testers, the targets config could be simplifed because
there was no longer a distinction between the tests in scope and all
tests.

Bug: 1240265
Change-Id: If4e75441a034aee7ed28fdfc95f69888626b50e9
Recipe-Manual-Change: chrome_release
Recipe-Manual-Change: build_limited
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3093174
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>