runtest.py: Add the JSON output arg to 'graphing' type tests.
crrev.com/c/3034599 added result_adapter to every type of gtest.

However, there's a super-special-case type of test on the
chromium.webrtc builders. Specifically browser_tests on those bots which
runs locally under a special execution mode of runtest.py

That execution mode is triggered when --annotate=graphing is passed to
runtest.py and cause the '--test-launcher-summary-output' arg to be
excluded, which makes result_adapter fail because it needs the json.

So this adds the '--test-launcher-summary-output' arg to any type of
runtest.py execution. This should hopefully be safe since:
- runtest.py is mostly unused, so blast-radius should be limited
- the output arg is pretty universal, so hopefully every type of test
  wrapped by runtest.py supports it

Bug: webrtc:13073
Change-Id: I393f4420873e94dcee4bc06c12d18c2dcfd837b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3100413
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>