Clean up isolate handling in tests.
Currently there are two properties for configuring isolate details of a
test: uses_isolate and isolate_target. These can be implemented
separately in unexpected ways: a test could indicate it does not use
isolate and return an isolate target (the result of the current default
implementations) or a test could indicate it uses isolate and not return
an isolate target.

This changes it so that the default for isolate_target is to return None
and Test subclasses that use isolate should override it to return a
sensible value. The implementation of uses_isolate returns whether or
not the isolate target is set to a non-false value and does not need to
be overridden in subclasses.

Change-Id: Ic8c2a4f0c4fc5559bf90c856448faee7d3d72410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3102583
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>