Don't customize gtest step results with test results in RDB experiment.
By default, a swarming gtest will modify the test step with subtext
regarding failures, which is populated via JSON.

With the switch to using RDB, we're trying to break the dependence on
parsing results JSON in the recipe. So this CL will remove all that
flavor text from a swarming gtest's step. Now, when a gtest fails, it
only turns red and lacks any info about which test cases failed.

The per-test info is still present in the 'test results' tab in the
new milo UI. And we can restore this info in the old build page via the
RDB results if desired. But for now, this leaves them absent.

Example LED build w/ this patch:
https://ci.chromium.org/swarming/task/5578a1a3c3cc3b10?server=chromium-swarm.appspot.com

Bug: 1135718
Change-Id: Id6a526625807cba29ff3bcee3297eebeb16d2854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3108007
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>