Dont customize step results for IsolatedScripts in the RDB experiment.
Similar to crrev.com/c/3108007, but for isolated script tests instead
of gtests.

This CL removes the step subtext describing how many tests passed,
flaked, and failed for for isolated script tests.

The per-test info is still present in the 'test results' tab in the
new milo UI. And we can restore this info in the old build page via the
RDB results if desired. But for now, this leaves them absent.

LED build w/ this patch:
https://ci.chromium.org/swarming/task/558db2e5cb71d010?server=chromium-swarm.appspot.com

Bug: 1135718
Change-Id: I671e7979c12b8b2fc38d16aaa8ec2c86ed43841f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3115305
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>