Restore uploading to test-results for swarmed tests in RDB experiment.
All builds in the RDB experiment today are skipping the step that
uploads all test results to the legacy 'test-results' service. This was
intentional given ResultDB's intended replacement of the legacy
test-tracking system.

As it turns out, that service is still used in production by find-it.
(See https://crbug.com/1243747.) So to preserve findit's ability to
detect and track flaky tests, we need to continue uploading to that
service for the foreseeable future.

This CL restores the uploading only for swarmed tests. This should be
"good enough" for most/all of findit's use-cases. But uploading of
local tests can be added in a follow-up if needed.

LED build w/ this patch applied:
https://ci.chromium.org/swarming/task/559d00c163472510?server=chromium-swarm.appspot.com
The upload step is hitting 409s from test-results, but that's due to
it being a led build. We'll see if the data has any trouble being
slurped up on builds with the experiment enabled when this lands.

Bug: 1135718
Change-Id: Ie2e5da8732bedc9052a9d63f4596d9642d6958e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3123659
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>