Allow a test's RDB results to update its exit_code after init.
In the current/legacy flow, the chromium recipe updates a test step's
display text with a list of failed test cases. eg: You can see that
display text rendered here in a legacy build: http://shortn/_Ux0JsyMzDH

But in the new world where we're not parsing JSON, that same step looks
like: http://shortn/_2Od93zCr1t
Note the lack of a list of "deterministic failures".

Today, that list can't be rendered for these new builds because it
requires modifying the "step_result" of the step representing that test.
And today we're fetching RDB results for the test in a separate RPC to
RDB after we render the step in question. And the recipe engine does
not allow modifying the presentation of any step other than the most
recent one.

So to allow RDB-based builds to display the same info in the same step,
we must run that RDB query RPC _before_ we render the display step.
That will be implemented in a follow-up CL.

But before we can do that, we must allow a RDBPerSuiteResults to update
its exit code after initialization. This is because the step display for
a swarming test is also the step that collects the exit codes of all the
swarming shards of that test, which leads to somewhat of a
chicken-and-egg problem. ie: We need the RDB results before we run
`collect`, but we need to run `collect` to fully initialize a test's
RDB results.

Adding this little re-init method should help avoid that.

Bug: 1135718, 1245085
Change-Id: Ib352f57d85a3c1733dfca6154c713ffb4400c865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3130350
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>