Implement experimental tests using the spec disabling.
For handling ci_only tests, a generic method of disabling test specs was
added. Implementing experimetal tests in terms of this allows for
simplifying some of the code. Because get_test won't be called for test
specs that are disabled, the returned test does not need to account for
it being disabled and test objects won't be passed around for disabled
test specs.

Change-Id: Idece26c23f09dde67625a5cb7264bf259bb924de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3133938
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>