Render swarming test steps with failure info in RDB experiment.
Similar to what we do for failures when not in the RDB experiment, this
will render a summary of failed test-case in a test's step. This
includes listing each failed test-case, along with a sub-log for each
one.

That sub-log would traditionally have a log snippet from the test's
output. However, that info currently isn't tracked in our RDB results.
Doing so is possible, but non-trivial as it appears to require a
separate RPC to RDB for each snippet.

So for now, the text in that sub-log will simply redirect the reader
to the "Test Results" tab in the new Milo UI.

LED build w/ this patch applied:
http://shortn/_SqJHQfq8k3
Example sub-log of a failed test:
http://shortn/_0i5909m0B4

Bug: 1135718, 1245085
Change-Id: Ic0fcea2fe99355620f858e51fbf6948a6df25348
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3137313
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>