Fix py3 incompatibilities in code that is run when gathering tests.
When any module is flipped to be python3-compatible, it causes the tests
to be gathered for the entire repo using python3. This change will fix
code that executes at that time so that modules can be updated to
support python3.

Change-Id: Ib875dc397e2d18e0cdafd66fefc0f93929c13b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3144539
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>