Fix reason for skipping (without patch) steps.
Currently when skipping the (without patch) steps, the reason will
always be indicated that it is due to the testing/buildbot files being
changed. This is not the case if the try builder has the
retry_without_patch attribute set to False.

This change updates it so that multiple reasons can be displayed,
correctly indicating that it is disabled on the builder config if that
is the case. Additionally, instead of having a fixed glob of all of the
.json files and .pyl files in testing/buildbot, it now checks if the
affected files includes source side spec files that the builder actually
uses. The .pyl files are never read by the builder, they're just the
inputs that the .json files are generated from, so they don't impact
execution of the builder.

Change-Id: I42e00109d558e5b84b1e4c7939bb531f980f98ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3150032
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>