Make schedule_suites_retry tolerant to exception repr changes.
A change has been made to the repr of exceptions to ease transition to
python3, but this test is looking for an exact match on the format.
Using a regex will allow the upstream change to be rolled in.

Change-Id: I843412f75171f48672f0106c31f709048e23868a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3154379
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Commit-Queue: Xinan Lin <linxinan@chromium.org>