Remove reliance of BuilderConfig implementation on TrySpec.
This modifies the BuilderConfig type so that it can be created without
having to create a TrySpec. Instead, the BuilderConfig will have the
fields that TrySpec has, with potentially different defaults.

The lookup method will look for a TrySpec in try_db and if it finds one,
it will use the values from the TrySpec for the shared fields.

This is a part of cleaning up the shape of the builder config so that it
can be exposed as a proto.

Bug: 1248219
Change-Id: I958cca13cc6344e7c02a4896c44982ef7f408a95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3150670
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>