Remove the TryDatabase from the builder_config.
The TryDatabase will now only be used during the lookup operation to
find the mirrored or mirroring builders. The only use of the try_db
attribute was to compute the try builders that mirror a builder. This is
now done as part of the lookup operation and exposed via the
mirroring_try_builders attribute.

This is a part of cleaning up the shape of the builder config so that it
can be exposed as a proto.

Bug: 1248219
Change-Id: Ic78da480812693fbfdc7e460cba6bc82f7de7483
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3157261
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>