Remove uses of active_result in chromium_tests/steps.py.
Using active_result in a finally block is error-prone, resulting in 2
possible issues if an exception is raised in the try scope before the
step is executed, depending on the state of the previous step. If the
previous step is closed, attempting to access active_result causes an
AssertionError to be raised, hiding the original exception. If the
previous step is not closed, the finally block will be accessing the
result of a potentially-unrelated step.

This changes the steps.py code to set raise_on_failure to False when
calling api.step() and then calling api.step.raise_on_failure() after
processing the result to trigger the appropriate exception. If an
exception is raised before the step is executed, code after the step
call won't be executed, so it won't hide errors or modify the incorrect
result. The status value is preserved before operating on the result so
that the exception type raised by raise_on_failure() will match what
would have been raised by step().

Bug: 1246180
Change-Id: If995861dd23d56f06b269af73c139c1ee2f2521a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3154689
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>