Remove the TryMirrors from the builder_config.
The TryMirrors will now only be used during the lookup operation to
determine the builder IDs and the builder IDs in scope for testing. The
only use of the mirrors attribute was for reporting the builders being
executed which now operates on the builder IDs.

This is a part of cleaning up the shape of the builder config so that it
can be exposed as a proto.

Bug: 1248219
Change-Id: I6b5f6f03b33a50fb825cf8afa979adb8d26232a6
Recipe-Nontrivial-Roll: build_limited
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3159246
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>