Add experiment to enable triggering children with GitilesTrigger.
The GitilesTrigger allows for specifying the gitiles commit. This
enables propagating the ref information so that the gitiles information
of the triggered builder can match that of the triggering builder
without having to write ad-hoc code.

This will allow the Android official builders to propagate a commit with
a ref so that the Android official testers can set the output commit.
Setting the output commit requires that the input commit either has a
ref or the checked out revision must have a position footer, but the
Android official yesyers run on commits that publish DEPS for the
version and do not contain a commit position.

Bug: 1249938
Change-Id: I77e2b13c53160fe5ca3b2cee3a2596c788bf3fb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3161484
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>