Only use gitiles trigger when in the experiment.
The intention of https://crrev.com/c/3165497 was that a commit could be
passed in that would be used only if the build is in the
use_gitiles_trigger experiment. The change mistakenly used the gitiles
trigger if the commit was provided even if the build was not in the
experiment.

Bug: 1249938
Change-Id: Ia35b938f5cc33903ec046034e4a9eabec71c39ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3171824
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>