Fix recipe test coverage in recipes/android/sdk_packager.py
The recipe test coverage of this recipe is weirdly flaky. See the
presubmit builds on https://crrev.com/c/3182890/2.

Sometimes the coverage passes, sometimes it doesn't.

That's assumed to be because the step_test_data here is initialized
once then re-used for multiple tests. And the order in which the tests
run may impact how they use/modify that test data.

So this tries to fix it by initializing the test data once for every
test.

And further qualify the tests that expect StatusException but also
asserting on their human-readable build status.

No-tree-checking since this should fix the tree.

No-Tree-Checks: true
Bug: 1252976
Change-Id: I06925ec4d54fda75fba429259cd2a64175df7ecc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3183029
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>