[Flakiness] Preparing tests for 20 executions
This CL updates the TestDefinition object to parse out the needed
fields from the ResultDB TestResult proto, which includes variants
and tags that contain test_suite and test_name information.

The test objects generated for a given builder are cross-referenced
with the test_id to find a corresponding Test object. The
TestOptions are overridden with a 0 retry count and 20 repeats.

These TestOptions still need to be translted into the corresponding
arguments for the MVP on iOS.

Bug: 1204163
Change-Id: I54f6bd9edd0ea2194950a7a449af05967f06b5c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3177290
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>